refactor(blog-client): :art: removed notion-utils package
