--- packages/notion-utils/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "@coldsurfers/notion-utils",
-  "version": "1.0.1-rc.10",
-  "main": "src/index.ts",
-  "source": "src/",
-  "files": [
-    "src",
-    "dist"
-  ],
-  "exports": {
-    ".": {
-      "import": "./src/index.ts",
-      "require": "./src/index.ts"
-    }
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@notionhq/client": "*",
-    "@types/sha1": "^1.1.5",
-    "axios": "*",
-    "eslint": "*",
-    "form-data": "^4.0.0",
-    "prettier": "*",
-    "release-it": "*",
-    "sha1": "^1.1.1",
-    "typescript": "*"
-  },
-  "peerDependencies": {
-    "@notionhq/client": "*",
-    "axios": "*",
-    "form-data": "^4.0.0",
-    "sha1": "^1.1.1"
-  },
-  "scripts": {
-    "prebuild": "yarn workspace @coldsurfers/cloudinary-utils build && yarn workspace @coldsurfers/shared-utils build",
-    "build": "rm -rf dist && tsc --project ./tsconfig.build.json",
-    "prerelease": "yarn build",
-    "prerelease:rc": "yarn build",
-    "release": "release-it minor --ci",
-    "release:rc": "release-it --ci --preRelease=rc",
-    "lint": "eslint --cache \"src/**/*.{js,jsx,ts,tsx}\"",
-    "lint:fix": "yarn lint --fix && prettier --write ."
-  },
-  "publishConfig": {
-    "registry": "https://npm.pkg.github.com"
-  },
-  "release-it": {
-    "git": {
-      "commitMessage": "chore: release @coldsurfers/notion-utils v${version}",
-      "tagName": "notion-utils-v${version}"
-    },
-    "github": {
-      "release": true
-    },
-    "hooks": {
-      "after:git:release": "git reset --hard HEAD"
-    }
-  },
-  "dependencies": {
-    "@coldsurfers/cloudinary-utils": "workspace:*",
-    "@coldsurfers/shared-utils": "workspace:*"
-  }
-}
