--- packages/notion-utils/src/index.ts
@@ -1,2 +1,0 @@
-export * from './utils'
-export { default as notionInstance } from './notionInstance'
