--- packages/notion-utils/src/notionInstance.ts
@@ -1,7 +1,0 @@
-import { Client } from '@notionhq/client'
-
-const notion = new Client({
-  auth: process.env.NOTION_TOKEN,
-})
-
-export default notion
