--- packages/notion-utils/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './notions'
