--- packages/notion-utils/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "noEmit": false,
-    "module": "CommonJS"
-  },
-  "include": ["."],
-  "exclude": ["node_modules"]
-}
